+++ CHANGELOG.rst
@@ -74,6 +74,7 @@ Version 2021-dev
 -  Refactored CUDA-OpenMP integration (#571)
 -  improved kmc output (#574)
 -  fix build on Ubuntu 16.04 (#577)
+-  use unique_ptr in factory (#581)
 
 Version 1.6.3 (released XX.08.20)
 =================================